refactor(cactus-web): define status types in a helper file & update their usage in field components
CACTUS-745
