+++ modules/cactus-web/src/FileInput/FileInput.tsx
@@ -490,7 +490,7 @@ const FileInfo: React.FC<FileInfoProps> = (props) => {
         )}
       </FileBox>
       {errorMsg && (
-        <StatusMessage id={statusId} status="error">
+        <StatusMessage id={statusId} status="error" className="status-message">
           {errorMsg}
         </StatusMessage>
       )}
@@ -672,7 +672,7 @@ const InnerFileInput = styled(FileInputBase).withConfig(
     }
   }
 
-  ${StatusMessage} {
+  .status-message {
     margin-top: 4px;
     align-self: flex-start;
   }
