+++ modules/cactus-web/src/Select/Select.tsx
@@ -20,9 +20,8 @@ import { omitMargins } from '../helpers/omit'
 import { positionDropDown, usePositioning } from '../helpers/positionPopover'
 import { isPurelyEqual, useMergedRefs } from '../helpers/react'
 import { useScrollTrap } from '../helpers/scroll'
-import { textFieldStatusMap } from '../helpers/status'
+import { Status, StatusPropType, textFieldStatusMap } from '../helpers/status'
 import { boxShadow, fontSize, isResponsiveTouchDevice, radius, textStyle } from '../helpers/theme'
-import { Status, StatusPropType } from '../StatusMessage/StatusMessage'
 import Tag from '../Tag/Tag'
 import TextButton from '../TextButton/TextButton'
